separate dbus interface module
